+++ DependencyInjection/LoremIpsumPermissionCheckerExtension.php
@@ -2,6 +2,7 @@
 
 namespace LoremIpsum\PermissionCheckerBundle\DependencyInjection;
 
+use LoremIpsum\PermissionCheckerBundle\PermissionChecker;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
@@ -13,5 +14,11 @@ class LoremIpsumPermissionCheckerExtension extends Extension
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yaml');
+
+        $configuration = new Configuration();
+        $config = $this->processConfiguration($configuration, $configs);
+
+        $definition = $container->getDefinition(PermissionChecker::class);
+        $definition->setArgument('$roles', $config['roles']);
     }
 }